Surprise: Merry Xmas, revised tweakdata.
Many of these are adjustments to formatting to ease future revision.
Others are increases in resistance, health, damage, or falloff in response to calls for greater challenge. 
Some may need to be tweaked back down based on feedback, but the game should be much less easy to solo with these.