fix broken team ai's piercing crew boost
Disabled some of goof's old code in order to allow Team Ai's AP Rounds crew boost to work again. 
Something is causing the cross() function for the Vector3 class to not... exist. Might be beardlib, idk, I couldn't figure it out. This means that NewNPCRaycastWeaponBase:_fire_raycast() will crash when using the vanilla function, which used cross() as part of its weapon spread calculations.
This fixed function is identical to the original except that it replaces that spread calculation with _get_spread(), bypassing cross().
Note to self: tag as at-risk for updates; vanilla updates to pd2 involving this function may cause CD to break since it's a copypasted version and not old-posthook-able