weapon data entry p2: Secondaries
IDK when I'll be "back". I might be done. I'm so fucking tired of the arguing and fighting, both here and in the resmod server. Especially in the resmod server. It's exhausting and makes me feel like a horrible person, and I just want to get along and for other people to be happy.

It's my fault really, and I'm sorry.