Bot shouting/marking improvements + idle reloads
Idle reload makes bots reload if their has 50% or less ammo in the mag when out of combat, this includes looped reloads (bot fired 2 shots = bot reloads 2 rounds, and so on, unless it's a single-reload weapon like the Garand). Needs testing with clients to check how the animation looks on their side, as I made it sync the animation start and end, but speed can't be synced with vanilla network code (and not syncing just makes them do a standard reload that won't match the weapon). That aside, it should be fine.

Civ intimidation was revamped, now the bots yell at civs as long as they're not on the ground, and as long as the civ has more than 10s of intimidation left (where they'll get up if left on their own while that "timer" runs out). They will also yell at escorts to make them move (and also yell at them from time to time when moving, for flavor).

Enemy domination was added to the bots, it can be fully or partially enabled, the latter means the bots will assist the players with enemies that already surrendered (set to this currently), while the former means they will do it on their own (they get really spammy, but I would recommend not using full domination in the first place).

In the two cases above, bots will turn around to face valid targets first, and won't yell at them if they're not in front of them (unless they're really close).

Enemy marking was changed so that the bots can mark specials, normal enemies and cameras during stealth (currently doesn't work in casing mode, but the code is functional), Turrets, and Tasers that are currently tasing them. Except for the Taser part, they will need to have these units in front of them to be able to mark them, instead of just requiring line of sight with their head from any angle (unless really close). Except when being tased, they will also not mark units that are already marked (unless if something like giving them the HVT skill is desired, in which case, they will mark and upgrade "lower-tier" contours for a better bonus).

For the three shouting mechanics I added here, the one thing that needs testing is, although minor, if the bots correctly don't use gestures when near targets without facing them, as I changed the angle value recently.

Also, enjoy the shit-load of comments I added to the file.