Crash prevention + disabling turn during stealth
The former is just a line that shouldn't even be there, or should have a "if current_attention then" check, but it's useless anyway. The latter needs more testing.

I suddenly had the game crash because of these two today, but not yesterday. Now the game doesn't crash but I'll keep testing tomorrow.