Fixes + removed useless comments
Fixed the tag issues, or at least, made sure they work.

Improved the AP ammo check so that it directly checks if the weapon used by the unit has AP ammo (will make a few changes to the weaponbase files).

Slight change to the revive call check, if carrying a bag that slows the bot down and they're not in range to use Inspire on the player, drop the bag. If they're in range but Inspire is on cooldown, also drop the bag. I think I modified this on my end due to other changes and forgot to change it back so it works correctly.