Flamethrower rework + proper suppression + FF fix
Now uses multiple cones to collect hits per fired shot, based on the amount of rays set in weapontweakdata. This makes it similar to single_damage_instance shotguns (vanilla). It collects hits and deals damage once per enemy, to avoid any unwanted extra damage.

It also makes more use of the spread stat to decide where the cones will go.

The reason for this is simple, the way it deals damage is now more accurate with the flame effect, instead of being near the center of it.