Modified the fix + removed redundant check
Better to check first if the attacker is a player and if they're alive, before checking the melee weapon. Sorry for the inconveniences, I already had this check on my end and thought I ported this when I modified all the functions, I guess I missed it with it being related to achievements.

Also updated a check I forgot to.. update before. Also removed an alive() check from the contour distance damage bonuses for both fire and explosion since that should be done after determining whether or not the attacking unit is a normal attacker or a throwable (in which case it points the thrower as the attacker).