Actual warp crash fix (for player husks)
While checking for the server should be done, as warping should only be done by them, the issue was that I put things in the wrong order so both in_air checks were done even for husks. This should also look a bit more clean.

Also had to use unit:position() for the player as player.pos doesn't get updated and causes issues.