Shotguns no longer almost guarantee shield_knock
Regardless of what system is used, damage per pellet or not, a Shield's shield only takes damage once and damage doesn't get reduced or increased. This puts shotguns on level with other weapons when it comes to knocking back shields with shots, as having up to 12 rolls is ridiculous, even if damage is split equally across pellets (which was done regardless of the system against shield objects).

Also added a better check for headshot kills, so that the specific case of the Headless Dozer's head hitbox not being actually disabled doesn't affect the achievement checks.