Flamethrower Rework - Fuck Cones Edition
Use a sphere ray to hit things instead of multiple broken cones, or a broken cone like in vanilla. This method is more clean as it properly gets direction, position and normal from the hits. It's also way better for performance, and it's actually very accurate when looking at the flames and what should get hit by them.

Melting a Dozer's faceplate/visor or a Turret's shield should be more consistent (and easier), as the previous method had a bit of RNG involved with the spread, even if it also had checks to prioritize those parts.