Added mod compatibility warning on startup (1/2)
Creates a QuickMenu popup on game startup warning the player if they have any mods (from a predetermined list), which might or will definitely conflict with Crackdown.
Mod installation is done by checking for the presence of a global value associated with that mod rather than scanning for folder names, since folder names may vary among installation versions of said mods. 
Feel free to add more mods to the list in this file, or just ask me to do it if you're can't do it at the moment or are not comfortable doing it. 
Currently listed mods: 
Better Bots
Full Speed Swarm
Iter
HoxHUD
Restoration Mod