More conditions to taunt after a charge kick
To avoid those cases where the victim got revived or something and the Cloaker still uses the voiceline when exiting the action. Since the voiceline plays for whoever's handling the attack locally (and gets synced to other players), this should always be accurate.

Also disabled the spawning of the baton on hand for flying kicks. It usually ends up on the Cloaker's back if the animation was sped up or when waiting to expire, and otherwise it's barely noticeable.