Ammo pickups are now partially shared with teammates (Total Crackdown)
Picking up an ammo pickup now also gives a portion of the ammo to teammates. The portion is: (normal pickup amount) / (number of players)
Also added some ammo pickup code to raycastweaponbase for when the skill overhaul portion of total cd (namely, the skill that gives an ammo pickup bonus to heavy weapons) is added.
The Heavy Weapons check is based on a boolean flag IS_HEAVY in weapontweakdata; flags are not yet implemented, though the check for them is.
Similarly, the check for the "heavy weapons ammo pickup rate bonus" skill is implemented, though the skill itself is not. (This should not affect the current function.)