Fixed Holdout not granting any EXP for clients
The issue was related to how the difficulty change was synced (which wasn't going to work properly when reaching CD difficulty anyway, due to the network function clamping the difficulty index to 8). So, instead, we'll make it happen locally for all peers.

Also made it so players dropping into a Holdout in progress get the difficulty change applied properly as well.