Hold grenade to preview tripmine, throwable crash fix
Holding down the [throw grenade] keybind to place a tripmine in TCD will now preview the tripmine placement, and place the tripmine upon releasing the key. The tripmine "dummy unit" preview will only appear if the placement is valid. (todo: on-screen text indicating placement validity)
The throwable equipment-override framework initially implemented remains in the codebase should it be needed in the future.
Fixed a crash when throwing a non-tripmine throwable