Implemented Static Defense functionality
Just need to add the upgrade and have it sync + set the flags for when that happens + add Static Defense's ace alert sound.

Also fixed two issues with the destroy function (even outside of TCD), like not calling the unit's base extension's destroy listener, and destroying the nav tracker used for sabotage SOs, if it was ever created. Both are very severe.