Fixed a client nav_link related crash
Could happen if the unit had to move from one nav_link to another with no standard nav_points in between, these tend to cause trouble in general, like with the previous crashes.

In this case, this was happening if a walk action expired for the server and the unit was still doing a nav_link animation on a client. A shortcut attempt is made when this happens once the animation is done (this normally isn't done with vanilla code, which can lead to the unit walking to the next nav_point and then walk back to where it stopped with the nav_link).