Updated unitnetworkhandler (TCD trip mines)
To work with the new enemy attaching code.

Plus, now, since a Body is also synced, that one has priority over the synced Object3D, as the latter don't sync properly (so its more effective to sync a body and get the root Object3D of that body, than use a synced Object3D). So, stuff like attaching a trip mine to a Dozer's leg, but other people will see it attached to an arm, or around the head, or floating in the air, etc, shouldn't happen anymore.