(TCD) Sentry Overheat mechanic, part 2
Removed sentry firemode interaction (previously used to open sentry radial menu)
Replaced "pickup" sentry interaction with radial menu interaction
Added "Retrieve" option to sentry radial menu, to pick up your sentry; the end result is that all sentry interactions are done via interacting directly with the sentry, with a 2m interaction radius (changing firemodes, changing ammo types, and retrieving sentries via the radial menu; opening the radial menu or venting the sentry heat via interacting with the sentry)
Removed "High Explosives" ammo option for sentries
Selecting the same ammo type as the ammo type that is currently equipped to a sentry via the radial menu no longer reverts the sentry to basic ammo
In contrast, selecting the same firemode that is currently equipped to a sentry via the radial menu now reverts the sentry to the default firemode
Removed a lot of old sentrycontrolmenu code, particularly that relating to multi-sentry selection capability
Fixed sentry interaction prompt text being shown when the sentry radial menu is open
Increased the opacity of radial sections "body" bitmap in the sentry radial menu (the thick arc slice in the inner circle). This section is now hidden or shown to indicate whether or not the relevant mode/ammo type is currently active on that sentry.