(TCD) Fixed ammo pickup bug
Addressed an issue where receiving shared ammo pickups could result in glitched ammo counts such as -1, which would cause weapon functionality issues.

The root cause of this issue is that GroupAIStateBase:num_alive_players() returns the size of the _player_criminals array using the # operator; however, the contents of _player_criminals are indexed by unit key (a userdata type), so this function will return 0. PAYDAY 2's table.size() implementation should instead be used to get the size of a non-sorted array, or any table with indices that are not sequential integers starting at 1.

This causes a div-0 value (inf) to be passed instead of the expected ammo ratio values of 1, 1/2, 1/3, or 1/4, resulting in the ammo issue.
The fix in this commit is a workaround to address this specific ammo sharing issue rather than changing the functionality of GroupAIStateBase:num_alive_players() itself, until the extent to which such a fix would affect other game mechanics may be ascertained.