(TCD) Added more specific feedback for Convert failure
Added slightly more specific hint feedback for when an attempt to convert an enemy fails (when using the Friendship Collar)
If you are the host, and your attempt fails due to having insufficient convert slots, the hint will display the current/max number of converts.
If you are not the host, and the host determines that your attempt fails due to insufficient convert slots, you will be shown a hint saying that there are too many converts.
If the unit you are trying to convert is already converted, you will be shown a hint saying that the unit is already converted.

Also fixed an inconsistent internal naming for a localization string