(TCD) The CSV Weapon Attachment update
Finished implementation of csv file based weapon attachment rebalancing. This will make it much easier to edit stats for weapons and attachments in the future.
Stat balancing via csv file has been enabled for both weapons (primary/secondary) and weapon attachments.
Added csv files for attachments and weapons. The spreadsheet document of weapons will be available for any user to copy and tweak to their liking at a later date, once the formatting is cleaned up a bit more.
Cleaned up the original weapon and attachment stat tweakdata files in weapontweakdata and weaponfactorytweakdata, now that 99% of weapon stat changes are done through the spreadsheet and parser.
All suppressor-related weapon attachments (any part marked with sub-type "silencer" or with perk "silencer") bestow maximum threat reduction and the "subclass_quiet" subclass.
Changes to attachment stat systems such as linear total ammo bonus (1:1, no complications)
Updated some blackmarketgui and playerinventorygui to accommodate changes to stat systems
Updated matchmaking key
Removed an easter egg such that John Wick is no longer a precision character (was a bug, fixed, then restored intentionally as an easter egg; removing now because it's being mistaken for a bug)
Squashed some bugs in the parser- internal threat calculation was incorrect in the spreadsheet document, so threat-to-suppression should now be done directly in the parser (at runtime). In the future, all such stats should be calculated in this way, including Total Ammo mul, RoF, accuracy, stability, price, and reload speed mul.

Please note that shotgun mods and ammo types have not been implemented in this commit; they will be added in a future update soon, as they are sufficiently different and complex to warrant a separate effort.

Known issue: Stat deltas for some stats (such as the new 0.5x Total Ammo multiplier with the Speedpull Mag, or the +5 Total Ammo addend for the Locomotive 12G's Shell Rack) will not display in the blackmarket preview, although the actual end stat will be displayed correctly. This will be addressed in a later update.